feat/#20-type 모듈 생성
