+++ src/components/Type.jsx
@@ -1,0 +1,34 @@
+import styled from "styled-components";
+
+const Types = ({ type }) => {
+  return <Type src={matchingFiles(type)}></Type>;
+};
+
+const matchingFiles = (type) => `/image/${TypeName[type]}.svg`;
+
+const TypeName = {
+  노말: "normal",
+  드래곤: "dragon",
+  불꽃: "fire",
+  물: "water",
+  전기: "electric",
+  풀: "grass",
+  얼음: "ice",
+  격투: "fighting",
+  독: "poison",
+  땅: "ground",
+  비행: "flying",
+  에스퍼: "psychic",
+  벌레: "bug",
+  바위: "rock",
+  고스트: "ghost",
+  강철: "steel",
+  악: "dark",
+  페어리: "fairy",
+};
+
+const Type = styled.img`
+  width: 60px;
+`;
+
+export default Type;
